requirements.txt
streamlit
sentence-transformers
faiss-cpu
llama-cpp-python
torch
pypdf